Created an scheduled service on backend to mark tickets as expired

+++ backend/src/main/java/com/tickethub/repositories/EventRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 import com.tickethub.domain.Event;
 
-public interface EventRepository extends JpaRepository<Event, Long>, JpaSpecificationExecutor<Event> {
+public interface EventRepository extends JpaRepository<Event, Integer>, JpaSpecificationExecutor<Event> {
 }
